added remaining methods in List and added BList
Added the remainder, unsure about the type casting of the set() function in the list, since we need to put an element wrapped in a node. Also for BList I used the super stuff, now we need to test the methods. Didn't address the commented methods in List yet, to review.